Add an `optString` property to TextField to support binding to optional-String-based properties.
This serves as a sort of workaround for the case where a list/tree cell is being deleted; the `optString`
property has special support for preserving the previous string content to avoid it disappearing abruptly
as the cell fades out.

Also, rename CellTextProperty -> TextProperty and centralize binding code.

Issue: VP-197