Supply initial value when deriving signal from relation rather than when creating property.
This makes it so that RelationSignal is aware of the initial value.  Previously, when an initial value was
supplied to the property, the signal wouldn't compute/deliver the initial value, but since its latest
value was left as nil, it would send a change even if the updated value was not actually different.