Fix an issue with AsyncReadableProperty.zip where zipped value would not be delivered if both underlying values were non-nil at init time.
This highlighted an issue with BinaryOpSignal (and other Signal implementations that rely on capturing latest value(s)
delivered by underlying signals), so for the time being those have been commented out, and BinaryOpValueProperty
has been rewritten to not rely on BinaryOpSignal so that we get the correct behavior for `zip`.  Eventually we
will find a better solution for Signals and initial value delivery.

Issue: VP-217