Checkpoint some initial refactoring of Signal and its change count stuff.
The original goal was to change AsyncReadableProperty so that it presents a public `signal`
that is distinct from the `underlyingSignal`, which allows us to guarantee that observers
of `signal` will be notified after `value` has been set.  But this minor change rocked the
boat and necessitated changes elsewhere, and now we need to fix up the fragile changeCount
and deliverInitial stuff.

Issue: VP-235