Checkpoint big refactoring of Signal and the property types.
This is very much a WIP, lots of things are broken and many tests haven't yet been updated.

The main idea here is to address longstanding weaknesses with change tracking (Signal.changeCount
is gone now, for example) and signal observation.  It is no longer necessary to explicitly start
a property or signal; instead, the underlying source of values is started when the first observer
is attached.  Additionally, observers receive the latest value when attached.

Issue: VP-235