Make specification of synchronous vs asynchronous signals a bit more rigid.
Synchronous signals are those that always have a value available immediately.  These signals
are expected to only deliver valueChanging events.  (This is pretty much how they used to behave;
I had recently changed them to deliver valueWill/DidChange immediately before/after but that
made things more complicated.)

Asynchronous signals may not have a value available.  There's now a SignalEvent enum that makes
the naming a bit more clear (beginPossibleAsyncChange instead of valueWillChange), although we
could stand to make the names better.  When observing an async signal, it should immediately
send either a valueChanging (if a value is known) or beginPossibleAsyncChange (if the initial
value needs to be computed asynchronously).

Also, change the synchronous ReadableProperty to extract its initial value from the underlying
signal (when `value` is accessed or the first time an observer is attached, to keep things
lazy) instead of having to provide both an explicit initial value and a signal.

Issue: VP-235